Refactor & add "utils" file for "TokenWrapper" contract
