+++ contracts/tokenwrapper/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod contract;
 pub mod state;
+pub mod utils;
 
 #[cfg(test)]
 pub mod mock_querier;
